edited attributes for city and place models, added second API request that returns each places' details

+++ db/migrate/20160222020912_create_cities.rb
@@ -2,8 +2,8 @@ class CreateCities < ActiveRecord::Migration
   def change
     create_table :cities do |t|
       t.string :name
-      t.integer :lat
-      t.integer :lng
+      t.decimal :lat
+      t.decimal :lng
       t.integer :radius
 
       t.timestamps null: false
